Replace "2.3-rc" alias with "2.3-dev"
Per a discussion with dcorbett (from HAProxy upstream), this changes our current `2.3-dev0` release to only have tags of `2.3-dev0` and `2.3-dev`.

We also discussed that `2.3-dev0` is not *super* meaningful for us to publish since it's the same as `2.2.0`, but it mirrors the versions on https://www.haproxy.org/ (and allows users to opt into testing these development snapshots early), and otherwise shouldn't cause any harm for us to publish.

As with any software, users who need a stable and reliable environment probably shouldn't deploy development snapshots/releases, but it's certainly useful to upstream to get users testing development builds early to help catch bugs before they hit GA.