Resolving missing ca-certificates dependency that causes client ssl verify to fail
Without this package, '/etc/ssl/certs' is either empty or incomplete, and
so does the @system-ca variable within haproxy.

This results in some haproxy client ssl features not working out of the
box. For instance, using httpclient with https endpoints will not work
with default config since ssl verify is on by default.

See PR #216.

Co-authored-by: Tianon Gravi <admwiggin@gmail.com>